Complete forest context api
